% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{AdjSigVarExp}
\alias{AdjSigVarExp}
\title{Adjust Signal Variation Explained}
\usage{
AdjSigVarExp(J, I, N, JntVarEx, IndVarEx)
}
\arguments{
\item{J}{joint signal matrix of size n-by-p}

\item{I}{individual signal matrix of size n-by-p}

\item{N}{noise/error matrix of size n-by-p}

\item{JntVarEx}{desired proportion of total variation explained by the joint signal}

\item{IndVarEx}{desired proportion of total variation explained by the individual signal}
}
\value{
a list of 3 items: 1) adjusted joint signal matrix; 2) adjusted individual signal matrix; 3) data matrix additively comprised of the adjusted signal matrices
}
\description{
Adjusts the proportion of total variation attributable to each signal component to predetermined values
}
