% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_CJIVE_Scripts.R,
%   R/SimulationStudyFunctions.R
\name{ConvSims_gg}
\alias{ConvSims_gg}
\title{Convert simulation study results}
\usage{
ConvSims_gg(AllSims)

ConvSims_gg(AllSims)
}
\arguments{
\item{AllSims}{matrix with each row representing results from a replicate in the simulation study described in CJIVE manuscript}
}
\value{
list of 2 items: 1) joint ranks determined by each method employed in the simulations study
                         2) chordal norms between true and estimated joint/individual loadings/scores for each method employed in the simulation study

list of 2 items: 1) joint ranks determined by each method employed in the simulations study
                         2) chordal norms between true and estimated joint/individual loadings/scores for each method employed in the simulation study
}
\description{
Convert results from simulation study into a form for graphing with ggplot

Convert results from simulation study into a form for graphing with ggplot
}
