% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{GetSimResults_Dir}
\alias{GetSimResults_Dir}
\title{Retrieve simulation results}
\usage{
GetSimResults_Dir(sim.dir, p1, p2, Preds = FALSE)
}
\arguments{
\item{sim.dir}{(character string) file path for the directory from which results will be retrieved}

\item{p1}{number of features in data set 1}

\item{p2}{number of features in data set 2}

\item{Preds}{(logical) do the replicate results contain correlations between predicted and true joint subject scores. Default is FALSE}
}
\value{
upper triangular p-by-p matrix
}
\description{
Retrives and compiles results from simulation study which are stored in a directory. A directory should contain separate .csv files (one per replicate),
 each of which will include all evaluation metrics and most experimental settings for that particular replicate. For the CJIVE manuscript, a directory houses results of all
 100 replicates for each combination of experimental factors.
}
