% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{MatVar}
\alias{MatVar}
\title{Matrix variation (i.e. Frobenius norm)}
\usage{
MatVar(X)
}
\arguments{
\item{X}{a matrix of any size}
}
\value{
The Frobenius norm of the matrix X, calculated as the square root of the sum of squared entries in X
}
\description{
Calculates the Frobenius norm of a matrix, which can be used as a measure of total variation
}
\examples{
X = matrix(rnorm(10), 5,2)
MatVar(X)
}
