% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_CJIVE_Scripts.R, R/Utilities.R
\name{MatVar2}
\alias{MatVar2}
\title{Matrix variation (i.e. Frobenius norm)}
\usage{
MatVar2(X)

MatVar2(X)
}
\arguments{
\item{X}{a matrix of any size}
}
\value{
The Frobenius norm of the matrix X, calculated as the sum of square entries in X

The Frobenius norm of the matrix X, calculated as the sum of square entries in X
}
\description{
Calculates the Frobenius norm of a matrix, which can be used as a measure of total variation

Calculates the Frobenius norm of a matrix, which can be used as a measure of total variation
}
\examples{
X = matrix(rnorm(10), 5,2)
MatVar2(X)
X = matrix(rnorm(10), 5,2)
MatVar2(X)
}
