% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{scale_loadings}
\alias{scale_loadings}
\title{Scale and sign-correct variable loadings to assist interpretation}
\usage{
scale_loadings(loading.comp)
}
\arguments{
\item{loading.comp}{numeric vector of variable loadings from a JIVE analysis}
}
\value{
numeric vector of loadings which have been scaled and sign-corrected
}
\description{
Scale loadings for a joint or individual component by its largest absolute value resulting in loadings between -1 and 1.
             Loadings are also sign-corrected to result in positive skewness
}
