% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{show.image.2}
\alias{show.image.2}
\title{Display a heatmap of a matrix (adapted from Erick Lock's show.image funciton in the r.jive package)}
\usage{
show.image.2(
  Image,
  ylab = "",
  xlab = "",
  net = F,
  main = "",
  sub = "",
  colorbar = TRUE
)
}
\arguments{
\item{Image}{matrix to display}

\item{ylab}{lab for y-axis of heatmap}

\item{xlab}{lab for x-axis of heatmap}

\item{net}{logical (TRUE/FALUSE) of whether entries correspond to edges between regions of interest in the Power-264 brain atlas. Default is FALSE}

\item{main}{main title for heatmap}

\item{sub}{subtitle for heatmap}

\item{colorbar}{logical (TRUE/FALUSE) of whether colorabar shouldl be included to aid interpretation. Default is TRUE}
}
\value{
graphical display of matrix as a heatmap
}
\description{
Visual display of a matrix as a heatmap with colors determined by entry values, and including a colorbar to aid interpretation of the heatmap
}
