% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{vec2net.l}
\alias{vec2net.l}
\title{Convert vector to network}
\usage{
vec2net.l(invector)
}
\arguments{
\item{invector}{numeric vector of size p choose 2}
}
\value{
lower triangular p-by-p matrix
}
\description{
Converts a vector of size p choose 2 into a p-by-p lower triangular matrix
}
