% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJIVE_Functions.R
\name{cc.jive.pred}
\alias{cc.jive.pred}
\title{CJIVE joint subject score prediction}
\usage{
cc.jive.pred(
  orig.dat.blocks,
  new.subjs,
  signal.ranks,
  cc.jive.loadings,
  can.cors
)
}
\arguments{
\item{orig.dat.blocks}{list of the two data matrices on which CJIVE was initially conducted}

\item{new.subjs}{list of two data matrices containing information on new subjects}

\item{signal.ranks}{a vector of length two which contains the rank for the signal within each data block. The rank corresponds to the number of principal
components (PCs) to be retained within each data block. If NULL, the ranks are determined by the parameter 'perc.var.' Default is NULL}

\item{cc.jive.loadings}{canonical loadings for the joint subspace}

\item{can.cors}{canonical correlations from the PCs of the data on which CJIVE was initially conducted - notated as rho_j in CJIVE manuscript}
}
\value{
matrix of joint subject score for new subjects
}
\description{
Predicts joint scores for new subjects based on CJIVE joint scores
}
