% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJIVE_Functions.R,
%   R/Functions_for_CJIVE_Scripts.R
\name{cc.jive.pred}
\alias{cc.jive.pred}
\title{CJIVE joint subject score prediction}
\usage{
cc.jive.pred(
  orig.dat.blocks,
  new.subjs,
  joint.rank = 1,
  signal.ranks,
  cc.jive.loadings
)

cc.jive.pred(
  orig.dat.blocks,
  new.subjs,
  joint.rank = 1,
  signal.ranks,
  cc.jive.loadings
)
}
\arguments{
\item{orig.dat.blocks}{list of the two data matrices on which CJIVE was initially conducted}

\item{new.subjs}{list of two data matrices containing information on new subjects}

\item{joint.rank}{The rank of the joint subspace i.e., number of components in the joint subspace}

\item{signal.ranks}{a vector of length two which contains the rank for the signal within each data block. The rank corresponds to the number of principal
components (PCs) to be retained within each data block. If NULL, the ranks are determined by the parameter 'perc.var.' Default is NULL}

\item{cc.jive.loadings}{canonical loadings for the joint subspace}
}
\value{
matrix of joint subject score for new subjects

matrix of joint subject score for new subjects
}
\description{
Predicts joint scores for new subjects based on CJIVE joint scores

Predicts joint scores for new subjects based on CJIVE joint scores
}
